+++ src/AuthenticatorInterface.php
@@ -11,7 +11,8 @@ interface AuthenticatorInterface
      * This is the class that communicates with the API and returns a JWT token.
      *
      * Any exception that it is thrown internally must be converted to a RequestException of guzzle.
-     * If the API returned a token but the token is not valid then a TokenRequestException must be thrown.
+     * If the API returned a token but the token is not valid then
+     * a {@see \Webthink\GuzzleJwt\Exception\BadTokenRequestException} must be thrown.
      *
      * @param string $username
      * @param string $password