+++ src/Exception/BadCredentialException.php
@@ -5,8 +5,6 @@ namespace Webthink\GuzzleJwt\Exception;
 use GuzzleHttp\Exception\RequestException;
 
 /**
- * Class BadCredentialException
- *
  * @author George Mponos <gmponos@gmail.com>
  */
 class BadCredentialException extends RequestException