+++ src/Exception/BadTokenRequestException.php
@@ -3,19 +3,16 @@
 namespace Webthink\GuzzleJwt\Exception;
 
 use Exception;
-use GuzzleHttp\Exception\BadResponseException;
+use GuzzleHttp\Exception\RequestException;
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
 
 /**
- * Class TokenRequestException
- *
  * @author George Mponos <gmponos@gmail.com>
  */
-class TokenRequestException extends BadResponseException
+class BadTokenRequestException extends RequestException
 {
     /**
-     * TokenRequestException constructor.
      *
      * @param string $message
      * @param \Psr\Http\Message\RequestInterface $request